+++ src/widget.js
@@ -266,6 +266,9 @@
                     title: experience.data.title,
                     context: 'mr2',
                     container: container,
+                    startPixel: config.startPixels && config.startPixels.join(' '),
+                    countPixel: config.countPixels && config.countPixels.join(' '),
+                    launchPixel: config.launchPixels && config.launchPixels.join(' '),
                     adId : adId
                 };
             }